Move KFold to validation.py
