+++ podium/validation/__init__.py
@@ -2,5 +2,4 @@
 Package contains modules used in model validation.
 """
 
-from .kfold import KFold
 from .validation import k_fold_classification_metrics, k_fold_validation, kfold_scores
